Omit R_ToplevelExec in README, mention R_tryCatch
After feedback from Luke:

I would encourage you to either not mention R_TopLevelExec at all or
explicitly discourage its use. It is intended for asynchronous things,
like handling events if that has to happen or running finalizers.
Things that would be in a different thread if they could be. In
particular code called from R_TopLevelExec sees different handler and
restart stacks from the point where R_TopLevelExec is called.

You could mention R_tryCatch, which provides a C variant of tryCatch.
Performance is not ideal since outsources the work back to R's
tryCatch, but that could in principle be improved eventually.